homeppp/requirements.txt
# 神经网络与深度学习课程设计 - 依赖包列表

# PyTorch核心库
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# 图像处理
Pillow>=9.0.0

# 数值计算
numpy>=1.21.0

# 可视化
matplotlib>=3.5.0

# 进度条
tqdm>=4.64.0

# 自然语言处理
nltk>=3.7

# 评估指标
rouge-score>=0.1.2
pycocoevalcap>=1.2

# 大模型微调（仅Model_finetune_Qwen2-VL需要）
transformers>=4.30.0
accelerate>=0.20.0
peft>=0.3.0

# 其他工具库
requests>=2.28.0

